LayoutTests/imported/w3c/web-platform-tests/shadow-dom/focus/focus-tabindex-order-shadow-varying-tabindex-3-expected.txt


FAIL Order with different tabindex on host assert_equals: expected Element node <button>Button</button> but got null
